--- src/main/java/codesquad/cafe/domain/article/repository/ArticleRepository.java
@@ -1,19 +1,0 @@
-package codesquad.cafe.domain.article.repository;
-
-import codesquad.cafe.domain.article.domain.Article;
-
-import java.util.List;
-
-public interface ArticleRepository {
-    void save(Article article, final String id);
-
-    List<Article> findAll();
-
-    Article findById(Long id);
-
-    String findWriterByUserId(String writerId);
-
-    void update(Article article);
-
-    void deletePostById(Long postId);
-}
