--- src/main/java/codesquad/cafe/domain/reply/dto/ReplyRequestDto.java
@@ -1,14 +1,0 @@
-package codesquad.cafe.domain.reply.dto;
-
-public class ReplyRequestDto {
-
-    private String replyContents;
-
-    public ReplyRequestDto(final String replyContents) {
-        this.replyContents = replyContents;
-    }
-
-    public String getReplyContents() {
-        return replyContents;
-    }
-}
