--- src/main/java/codesquad/cafe/domain/user/dto/UserLoginRequestDto.java
@@ -1,19 +1,0 @@
-package codesquad.cafe.domain.user.dto;
-
-public class UserLoginRequestDto {
-    private String userId;
-    private String password;
-
-    public UserLoginRequestDto(final String userId, final String password) {
-        this.userId = userId;
-        this.password = password;
-    }
-
-    public String getUserId() {
-        return userId;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-}
