+++ src/main/java/codesquad/cafe/reply/dto/ReplyRequestDto.java
@@ -1,0 +1,14 @@
+package codesquad.cafe.reply.dto;
+
+public class ReplyRequestDto {
+
+    private String replyContents;
+
+    public ReplyRequestDto(final String replyContents) {
+        this.replyContents = replyContents;
+    }
+
+    public String getReplyContents() {
+        return replyContents;
+    }
+}
